requirements.txt
pygame
neat-python
numpy